Rename `alloc` to `allocator`
