+++ wasm-node/rust/src/allocator.rs
